refacto constants
