+++ lib/constants/canada.ts
@@ -1,1 +1,27 @@
 export const CLIENT_ORIGIN = 'SPA';
+export const CA_BASE_URL = 'https://mybluelink.ca';
+
+export const CA_ENDPOINTS = {
+    login: `${CA_BASE_URL}/tods/api/lgn`,
+    logout: `${CA_BASE_URL}/tods/api/lgout`,
+    // Account
+    myAccount: `${CA_BASE_URL}/tods/api/acctinfo`,
+    nextService: `${CA_BASE_URL}/tods/api/nxtsvc`,
+    preferedDealer: `${CA_BASE_URL}/tods/api/gtprfrdlr`,
+    // Vehicle
+    vehicleList: `${CA_BASE_URL}/tods/api/vhcllst`,
+    vehicleInfo: `${CA_BASE_URL}/tods/api/sltvhcl`,
+    status: `${CA_BASE_URL}/tods/api/lstvhclsts`,
+    remoteStatus: `${CA_BASE_URL}/tods/api/rltmvhclsts`,
+    // Car commands with preauth (PIN)
+    lock: `${CA_BASE_URL}/tods/api/drlck`,
+    unlock: `${CA_BASE_URL}/tods/api/drulck`,
+    start: `${CA_BASE_URL}/tods/api/evc/rfon`,
+    stop: `${CA_BASE_URL}/tods/api/evc/rfoff`,
+    locate: `${CA_BASE_URL}/tods/api/fndmcr`,
+    hornlight: `${CA_BASE_URL}/tods/api/hornlight`,
+    // System
+    verifyAccountToken: `${CA_BASE_URL}/tods/api/vrfyacctkn`,
+    verifyPin: `${CA_BASE_URL}/tods/api/vrfypin`,
+    verifyToken: `${CA_BASE_URL}/tods/api/vrfytnc`,
+};
